requirements.txt
apscheduler==3.10.4
dj-database-url==2.1.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
langchain==0.0.350
openai==1.5.0
pgvector==0.2.4
pip-chill==1.0.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
tiktoken==0.5.2
twilio==8.11.0
